# agent/openapi_spec.yaml
openapi: 3.0.0
info:
  title: SENTINEL API
  description: >
    The Business Flight Recorder — records every business decision with its full
    Fivetran data context, then traces what happened next using Gemini 3 causal AI.
    Powered by Fivetran MCP, Google Cloud Agent Builder, Gemini 3, BigQuery, MongoDB Atlas.
  version: 2.0.0
servers:
  - url: https://sentinel-38381883054.us-central1.run.app
    description: Production (Cloud Run)
  - url: http://localhost:8100
    description: Local development
paths:
...
  /api/ask/:
    post:
      operationId: ask_sentinel
      summary: Ask SENTINEL a question about the decision log
      description: >
        Answer questions using Gemini 3 with full scenario context:
        "Why did we raise prices?", "What caused the churn spike?",
        "What data existed when we made the hiring decision?"
        "Explain the Pearson correlation score."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question: { type: string }
                demo_scenario: { type: string, enum: [acmesaas, qwikster] }
      responses:
        '200':
          description: Answer from SENTINEL
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer: { type: string }
                  confidence: { type: number }
                  sources: { type: array, items: { type: string } }
                  relevant_decision_ids: { type: array, items: { type: string } }
  /api/agent/chat:
    post:
      operationId: agent_chat
      summary: Run the SENTINEL ADK agent (Gemini 3 FunctionTools — full multi-step orchestration)
      description: >
        Invokes the Google ADK code-first agent. The agent autonomously calls
        list_fivetran_connectors → trigger_fivetran_sync → get_metrics_snapshot
        before answering. Returns the final synthesized response after all tool calls complete.
        Slower than /api/ask/ but performs full causal chain reasoning with real-time data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string }
                session_id: { type: string, description: "Conversation session ID (auto-generated if omitted)" }
                demo_scenario: { type: string, enum: [acmesaas, qwikster] }
      responses:
        '200':
          description: Agent response after multi-step tool orchestration
          content:
            application/json:
              schema:
                type: object
                properties:
                  response: { type: string }
                  session_id: { type: string }
                  agent: { type: string, example: "SENTINEL-ADK" }
                  model: { type: string, example: "gemini-3-flash-preview" }
...